chore: refactor and move to internal (#210)
