+++ pkg/internal/common/helper.go
@@ -35,6 +35,33 @@ import (
 	"github.com/fatih/color"
 )
 
+var ChainMetadataMap = map[int64]types.ChainMetadata{
+	MainnetChainId: {
+		BlockExplorerUrl:            "https://etherscan.io/tx",
+		ELDelegationManagerAddress:  "0x39053D51B77DC0d36036Fc1fCc8Cb819df8Ef37A",
+		ELAVSDirectoryAddress:       "0x135dda560e946695d6f155dacafc6f1f25c1f5af",
+		ELRewardsCoordinatorAddress: "0x7750d328b314EfFa365A0402CcfD489B80B0adda",
+		WebAppUrl:                   "https://app.eigenlayer.xyz/operator",
+		ProofStoreBaseURL:           "https://eigenlabs-rewards-mainnet-ethereum.s3.amazonaws.com",
+	},
+	HoleskyChainId: {
+		BlockExplorerUrl:            "https://holesky.etherscan.io/tx",
+		ELDelegationManagerAddress:  "0xA44151489861Fe9e3055d95adC98FbD462B948e7",
+		ELAVSDirectoryAddress:       "0x055733000064333CaDDbC92763c58BF0192fFeBf",
+		ELRewardsCoordinatorAddress: "0xAcc1fb458a1317E886dB376Fc8141540537E68fE",
+		WebAppUrl:                   "https://holesky.eigenlayer.xyz/operator",
+		ProofStoreBaseURL:           "https://eigenlabs-rewards-testnet-holesky.s3.amazonaws.com",
+	},
+	AnvilChainId: {
+		BlockExplorerUrl:            "",
+		ELDelegationManagerAddress:  "0xDc64a140Aa3E981100a9becA4E685f962f0cF6C9",
+		ELAVSDirectoryAddress:       "0x0165878A594ca255338adfa4d48449f69242Eb8F",
+		ELRewardsCoordinatorAddress: "0x610178dA211FEF7D417bC0e6FeD39F05609AD788",
+		WebAppUrl:                   "",
+		ProofStoreBaseURL:           "",
+	},
+}
+
 func PrintRegistrationInfo(txHash string, operatorAddress common.Address, chainId *big.Int) {
 	fmt.Println()
 	fmt.Println(strings.Repeat("-", 100))
@@ -255,13 +282,13 @@ func ReadConfigFile(path string) (*types.OperatorConfig, error) {
 		return nil, err
 	}
 
-	elAVSDirectoryAddress, err := getAVSDirectoryAddress(operatorCfg.ChainId)
+	elAVSDirectoryAddress, err := GetAVSDirectoryAddress(&operatorCfg.ChainId)
 	if err != nil {
 		return nil, err
 	}
 	operatorCfg.ELAVSDirectoryAddress = elAVSDirectoryAddress
 
-	elRewardsCoordinatorAddress, err := getRewardCoordinatorAddress(operatorCfg.ChainId)
+	elRewardsCoordinatorAddress, err := GetRewardCoordinatorAddress(&operatorCfg.ChainId)
 	if err != nil {
 		return nil, err
 	}
@@ -270,20 +297,9 @@ func ReadConfigFile(path string) (*types.OperatorConfig, error) {
 	return &operatorCfg, nil
 }
 
-func getAVSDirectoryAddress(chainID big.Int) (string, error) {
+func GetRewardCoordinatorAddress(chainID *big.Int) (string, error) {
 	chainIDInt := chainID.Int64()
-	chainMetadata, ok := utils.ChainMetadataMap[chainIDInt]
-	if !ok {
-		return "", fmt.Errorf("chain ID %d not supported", chainIDInt)
-	} else {
-		return chainMetadata.ELAVSDirectoryAddress, nil
-	}
-}
-
-// TODO(shrimalmadhur): remove this and use the utils one in a separate PR
-func getRewardCoordinatorAddress(chainID big.Int) (string, error) {
-	chainIDInt := chainID.Int64()
-	chainMetadata, ok := utils.ChainMetadataMap[chainIDInt]
+	chainMetadata, ok := ChainMetadataMap[chainIDInt]
 	if !ok {
 		return "", fmt.Errorf("chain ID %d not supported", chainIDInt)
 	} else {
@@ -291,9 +307,19 @@ func getRewardCoordinatorAddress(chainID big.Int) (string, error) {
 	}
 }
 
+func GetAVSDirectoryAddress(chainID *big.Int) (string, error) {
+	chainIDInt := chainID.Int64()
+	chainMetadata, ok := ChainMetadataMap[chainIDInt]
+	if !ok {
+		return "", fmt.Errorf("chain ID %d not supported", chainIDInt)
+	} else {
+		return chainMetadata.ELAVSDirectoryAddress, nil
+	}
+}
+
 func GetTransactionLink(txHash string, chainId *big.Int) string {
 	chainIDInt := chainId.Int64()
-	chainMetadata, ok := utils.ChainMetadataMap[chainIDInt]
+	chainMetadata, ok := ChainMetadataMap[chainIDInt]
 	if !ok {
 		return txHash
 	} else {
@@ -303,7 +329,7 @@ func GetTransactionLink(txHash string, chainId *big.Int) string {
 
 func getWebAppLink(operatorAddress common.Address, chainId *big.Int) string {
 	chainIDInt := chainId.Int64()
-	chainMetadata, ok := utils.ChainMetadataMap[chainIDInt]
+	chainMetadata, ok := ChainMetadataMap[chainIDInt]
 	if !ok {
 		return ""
 	} else {
